+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.0.0
 
 ### New Features
-
- - The Formula Parser
+ - add new style `loop` which takes an array object and loop the rows and load its children passing the values of the rows
+ - The Formula Parser  based on [MathExecutor](https://github.com/neonxp/MathExecutor)